+++ packages/desktop/src/renderer/sagas/modals/modals.slice.ts
@@ -35,7 +35,9 @@ export class ModalsInitialState {
   [ModalName.loadingPanel] = { open: true, args: {} }; // Loading modal is open by default and closes on websocket connection
   [ModalName.channelCreationModal] = { open: false, args: {} };
   [ModalName.unregisteredUsernameModal] = { open: false, args: {} };
-  [ModalName.duplicatedUsernameModal] = { open: false, args: {} }
+  [ModalName.duplicatedUsernameModal] = { open: false, args: {} };
+  [ModalName.usernameTakenModal] = { open: false, args: {} };
+  [ModalName.possibleImpersonationAttackModal] = { open: false, args: {} }
 }
 
 export const modalsSlice = createSlice({