+++ packages/mobile/src/route.params.ts
@@ -42,6 +42,9 @@ export type RootStackParamList = {
   [ScreenNames.UnregisteredUsernameScreen]: {
     username: string
   }
+  [ScreenNames.UsernameTakenScreen]: undefined
+  [ScreenNames.NewUsernameRequestedScreen]: undefined
+  [ScreenNames.PossibleImpersonationAttackScreen]: undefined
 }
 
 export type SplashRouteProp = RouteProp<RootStackParamList, ScreenNames.SplashScreen>
@@ -59,3 +62,11 @@ export type ErrorRouteProp = RouteProp<RootStackParamList, ScreenNames.ErrorScre
 export type DuplicatedUsernameRouteProps = RouteProp<RootStackParamList, ScreenNames.DuplicatedUsernameScreen>
 
 export type UnregisteredUsernameRouteProps = RouteProp<RootStackParamList, ScreenNames.UnregisteredUsernameScreen>
+
+export type UsernameTakenRouteProps = RouteProp<RootStackParamList, ScreenNames.UsernameTakenScreen>
+
+export type NewUsernameRequestedRouteProps = RouteProp<RootStackParamList, ScreenNames.NewUsernameRequestedScreen>
+export type PossibleImpersonationAttackRouteProps = RouteProp<
+  RootStackParamList,
+  ScreenNames.PossibleImpersonationAttackScreen
+>