+++ packages/mobile/src/screens/Splash/Splash.screen.tsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux'
 import { initSelectors } from '../../store/init/init.selectors'
 import { initActions } from '../../store/init/init.slice'
 import { SplashScreenProps } from './Splash.types'
-import { Loading } from '../../components/Loading/Loading.component'
+import { Splash } from '../../components/Splash/Splash.component'
 
 export const SplashScreen: FC<SplashScreenProps> = ({ route }) => {
   const dispatch = useDispatch()
@@ -21,5 +21,5 @@ export const SplashScreen: FC<SplashScreenProps> = ({ route }) => {
     }
   }, [ready, route.params?.code])
 
-  return <Loading />
+  return <Splash />
 }