+++ CHANGELOG.md
@@ -1,4 +1,37 @@
-[quiet@2.0.0-alpha.10]
+
+[2.0.1-alpha.2]
+
+* UI layer for taken usernames for desktop and mobile
+
+* Change nickname for taken username
+
+* Map messages sent before changing username
+
+* Update registrar service to match new registration flow.
+
+* Add possible impersonation attack UI for desktop and mobile
+
+* Fix truncated long messages in channelInput component
+
+* Unblock mobile e2e tests
+
+* Prettify loading component on Chat screen (mobile)
+
+* Running Chromatic tests for forked PRs
+
+* Added e2e test for user joining community when owner is offline. Improved e2e tests
+
+* Bump github actions/* to versions using node16
+
+* Project can now be bootstraped on Windows (powershell)
+
+* Placeholder(...) for community name
+
+* No unregistered/duplicated label for system messages
+
+[2.0.0-alpha.11]
+
+* Customize Launch Screen on iOS
 
 * Suspends certain websocket events until backend becomes fully operative (faster and dumber frontend).