+++ packages/desktop/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# 2.0.0-alpha.18 (2023-10-04)
+## 2.0.1-alpha.2 (2023-10-09)
 
 
 ### Bug Fixes
@@ -14,6 +14,38 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ### Reverts
 
 * Revert "Remove afterAllArtifactBuild for linux" ([975d0df](https://github.com/ZbayApp/monorepo/commit/975d0df58494bdfba1270f6845152af4969e77ea))
+
+
+
+
+
+## [2.0.1-alpha.1](https://github.com/ZbayApp/monorepo/compare/quiet@2.0.1-alpha.0...quiet@2.0.1-alpha.1) (2023-09-25)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+## [2.0.1-alpha.0](https://github.com/ZbayApp/monorepo/compare/quiet@2.0.0-alpha.12...quiet@2.0.1-alpha.0) (2023-09-25)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [2.0.0-alpha.12](https://github.com/TryQuiet/quiet/compare/quiet@2.0.0-alpha.11...quiet@2.0.0-alpha.12) (2023-09-20)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [2.0.0-alpha.11](https://github.com/TryQuiet/quiet/compare/quiet@2.0.0-alpha.10...quiet@2.0.0-alpha.11) (2023-09-19)
+
+**Note:** Version bump only for package quiet