+++ packages/logger/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.1-alpha.2](https://github.com/TryQuiet/quiet/compare/@quiet/logger@2.0.1-alpha.1...@quiet/logger@2.0.1-alpha.2) (2023-10-09)
+
+**Note:** Version bump only for package @quiet/logger
+
+
+
+
+
+## [2.0.1-alpha.1](https://github.com/ZbayApp/monorepo/compare/@quiet/logger@2.0.1-alpha.0...@quiet/logger@2.0.1-alpha.1) (2023-09-25)
+
+**Note:** Version bump only for package @quiet/logger
+
+
+
+
+
+## [2.0.1-alpha.0](https://github.com/ZbayApp/monorepo/compare/@quiet/logger@1.7.0-alpha.1...@quiet/logger@2.0.1-alpha.0) (2023-09-25)
+
+**Note:** Version bump only for package @quiet/logger
+
+
+
+
+
 # [2.0.0-alpha.18](https://github.com/ZbayApp/monorepo/compare/@quiet/logger@1.7.0-alpha.1...@quiet/logger@2.0.0-alpha.18) (2023-10-04)
 
 **Note:** Version bump only for package @quiet/logger