+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.1-alpha.2](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.8...@quiet/mobile@2.0.1-alpha.2) (2023-10-09)
+
+
+### Bug Fixes
+
+* typo in function name ([7e00db1](https://github.com/TryQuiet/quiet/commit/7e00db1eb8d868cfe6682a1fa75f7e90b3a496bb))
+
+
+### Reverts
+
+* Revert "Match designs for mobile chat input (#1841)" ([e57fbca](https://github.com/TryQuiet/quiet/commit/e57fbcadcc51995c0360ee3b7b7fba201f4a9605)), closes [#1841](https://github.com/TryQuiet/quiet/issues/1841)
+
+
+
+
+
+## [2.0.1-alpha.1](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@2.0.1-alpha.0...@quiet/mobile@2.0.1-alpha.1) (2023-09-25)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [2.0.1-alpha.0](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@2.0.0-alpha.15...@quiet/mobile@2.0.1-alpha.0) (2023-09-25)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
 # [2.0.0-alpha.21](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@2.0.0-alpha.20...@quiet/mobile@2.0.0-alpha.21) (2023-10-09)
 
 **Note:** Version bump only for package @quiet/mobile
@@ -29,10 +61,47 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [2.0.0-alpha.18](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@1.10.8...@quiet/mobile@2.0.0-alpha.18) (2023-10-04)
 
+**Note:** Version bump only for package @quiet/mobile
 
-### Bug Fixes
 
-* typo in function name ([7e00db1](https://github.com/ZbayApp/monorepo/commit/7e00db1eb8d868cfe6682a1fa75f7e90b3a496bb))
+
+
+
+# [2.0.0-alpha.15](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@2.0.0-alpha.14...@quiet/mobile@2.0.0-alpha.15) (2023-09-22)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+# [2.0.0-alpha.14](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@2.0.0-alpha.13...@quiet/mobile@2.0.0-alpha.14) (2023-09-22)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+# [2.0.0-alpha.13](https://github.com/ZbayApp/monorepo/compare/@quiet/mobile@2.0.0-alpha.12...@quiet/mobile@2.0.0-alpha.13) (2023-09-22)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+# [2.0.0-alpha.12](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@2.0.0-alpha.11...@quiet/mobile@2.0.0-alpha.12) (2023-09-20)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+# [2.0.0-alpha.11](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@2.0.0-alpha.10...@quiet/mobile@2.0.0-alpha.11) (2023-09-19)
+
+**Note:** Version bump only for package @quiet/mobile
 
 
 
@@ -48,10 +117,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [2.0.0-alpha.9](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.5...@quiet/mobile@2.0.0-alpha.9) (2023-09-18)
 
-
-### Bug Fixes
-
-* typo in function name ([7e00db1](https://github.com/TryQuiet/quiet/commit/7e00db1eb8d868cfe6682a1fa75f7e90b3a496bb))
+**Note:** Version bump only for package @quiet/mobile
 
 
 
@@ -59,10 +125,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [2.0.0-alpha.8](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@2.0.0-alpha.7...@quiet/mobile@2.0.0-alpha.8) (2023-09-14)
 
-
-### Bug Fixes
-
-* typo in function name ([7e00db1](https://github.com/TryQuiet/quiet/commit/7e00db1eb8d868cfe6682a1fa75f7e90b3a496bb))
+**Note:** Version bump only for package @quiet/mobile
 
 
 
@@ -76,40 +139,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-## [1.10.8](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.7...@quiet/mobile@1.10.8) (2023-10-03)
-
-**Note:** Version bump only for package @quiet/mobile
-
-
-
-
-
-## [1.10.7](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.6...@quiet/mobile@1.10.7) (2023-10-03)
-
-**Note:** Version bump only for package @quiet/mobile
-
-
-
-
-
-## [1.10.6](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.5...@quiet/mobile@1.10.6) (2023-09-25)
-
-**Note:** Version bump only for package @quiet/mobile
-
-
-
-
-
-## [1.10.5](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.4...@quiet/mobile@1.10.5) (2023-09-15)
-
-**Note:** Version bump only for package @quiet/mobile
-
-
-
-
-
 # [2.0.0-alpha.6](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@2.0.0-alpha.5...@quiet/mobile@2.0.0-alpha.6) (2023-09-07)
-## [1.10.4](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.3...@quiet/mobile@1.10.4) (2023-09-15)
 
 **Note:** Version bump only for package @quiet/mobile
 
@@ -118,7 +148,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 # [2.0.0-alpha.5](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@2.0.0-alpha.4...@quiet/mobile@2.0.0-alpha.5) (2023-09-06)
-## [1.10.3](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.2...@quiet/mobile@1.10.3) (2023-09-15)
 
 **Note:** Version bump only for package @quiet/mobile
 
@@ -127,7 +156,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 # [2.0.0-alpha.4](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@2.0.0-alpha.3...@quiet/mobile@2.0.0-alpha.4) (2023-09-06)
-## [1.10.2](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.1...@quiet/mobile@1.10.2) (2023-09-15)
 
 **Note:** Version bump only for package @quiet/mobile
 
@@ -168,6 +196,69 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 # [1.11.0-alpha.0](/compare/@quiet/mobile@1.10.0...@quiet/mobile@1.11.0-alpha.0) (2023-08-29)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.10.8](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.7...@quiet/mobile@1.10.8) (2023-10-03)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.10.7](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.6...@quiet/mobile@1.10.7) (2023-10-03)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.10.6](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.5...@quiet/mobile@1.10.6) (2023-09-25)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.10.5](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.4...@quiet/mobile@1.10.5) (2023-09-15)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.10.4](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.3...@quiet/mobile@1.10.4) (2023-09-15)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.10.3](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.2...@quiet/mobile@1.10.3) (2023-09-15)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
+## [1.10.2](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.1...@quiet/mobile@1.10.2) (2023-09-15)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
 ## [1.10.1](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@1.10.0...@quiet/mobile@1.10.1) (2023-09-15)
 
 **Note:** Version bump only for package @quiet/mobile